+++ worlds/pokemon_bw/__init__.py
@@ -166,6 +166,7 @@ class PokemonBWWorld(World):
         self.encounter_by_method: dict[str, list[int]] = {}
         self.trade_data: dict[str, tuple[int, int]] = {}
         self.dexsanity_numbers: list[int] = []
+        self.seensanity_numbers: list[int] = []
         self.regions: dict[str, Region] | None = None
         self.region_distances: dict[str, int] | None = None
         self.max_distance: int = 0
@@ -408,6 +409,7 @@ class PokemonBWWorld(World):
                 "studio_castelia_type": self.studio_castelia_type,
                 "driftveil_random_move_id": self.driftveil_random_move_id,
                 "other_locations_species": self.other_locations_species,
+                "seensanity_numbers": self.seensanity_numbers,
             }
         return part